requirements.txt
# MCP Server Dependencies for SQL Agent
# Python 3.10+ required

# Core MCP Server - Using FastMCP for cloud deployment
fastmcp>=0.1.0

# Database drivers and ORMs
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.23

# Environment and configuration
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Async support
asyncio>=3.4.3
aiofiles>=23.2.1

# Utilities
typing-extensions>=4.8.0

# Optional: For enhanced features
# pydantic>=2.5.0  # For data validation
# uvloop>=0.19.0  # For better async performance on Linux/macOS

